limit akses role

+++ resources/views/pages/kategori-obat/index.blade.php
@@ -19,15 +19,19 @@
                 <div class="col-12">
                     <div class="card">
                         <div class="card-body">
-                            <a type="button" class="btn btn-primary" data-toggle="modal" data-target="#formCreate"><i
-                                    class="fa fa-plus"></i> Tambah</a>
-                            @include('pages.kategori-obat.create')
+                            @if (request()->session()->get('user')['role'] === 'APOTEKER')
+                                <a type="button" class="btn btn-primary" data-toggle="modal" data-target="#formCreate"><i
+                                        class="fa fa-plus"></i> Tambah</a>
+                                @include('pages.kategori-obat.create')
+                            @endif
                             <table id="defaultTable" class="table table-bordered table-striped">
                                 <thead>
                                     <tr>
                                         <th>No</th>
                                         <th>Nama</th>
-                                        <th>Aksi</th>
+                                        @if (request()->session()->get('user')['role'] === 'APOTEKER')
+                                            <th>Aksi</th>
+                                        @endif
                                     </tr>
                                 </thead>
                                 <tbody>
@@ -38,40 +42,42 @@
                                                 {{ $i }}
                                             </td>
                                             <td>{{ $item->nama }}</td>
-                                            <td>
-                                                <form id="formDelete{{ $item->id }}"
-                                                    action="{{ route('kategori-obat.destroy', $item->id) }}" method="POST"
-                                                    class="d-inline">
-                                                    @csrf
-                                                    @method('delete')
-                                                    <a type="button" class="btn btn-danger"
-                                                        onclick="handleDelete({{ $item->id }})">
-                                                        <i class="fa fa-trash" title="Hapus Data User"></i>
-                                                    </a>
-                                                </form>
+                                            @if (request()->session()->get('user')['role'] === 'APOTEKER')
+                                                <td>
+                                                    <form id="formDelete{{ $item->id }}"
+                                                        action="{{ route('kategori-obat.destroy', $item->id) }}"
+                                                        method="POST" class="d-inline">
+                                                        @csrf
+                                                        @method('delete')
+                                                        <a type="button" class="btn btn-danger"
+                                                            onclick="handleDelete({{ $item->id }})">
+                                                            <i class="fa fa-trash" title="Hapus Data User"></i>
+                                                        </a>
+                                                    </form>
 
-                                                <script>
-                                                    function handleDelete(id) {
-                                                        Swal.fire({
-                                                            title: 'Apakah kamu yakin?',
-                                                            text: "kamu akan menghapus data ini!",
-                                                            icon: 'warning',
-                                                            showCancelButton: true,
-                                                            confirmButtonColor: '#3085d6',
-                                                            cancelButtonColor: '#d33',
-                                                            confirmButtonText: 'Ya, hapus!'
-                                                        }).then((result) => {
-                                                            if (result.isConfirmed) {
-                                                                document.getElementById('formDelete' + id).submit();
-                                                            }
-                                                        })
-                                                    }
-                                                </script>
-                                                <a type="button" class="btn btn-warning" data-toggle="modal"
-                                                    data-target="#formUpdate{{ $item->id }}">
-                                                    <i class="fa fa-edit" title="Ubah Data User"></i>
-                                                </a>
-                                            </td>
+                                                    <script>
+                                                        function handleDelete(id) {
+                                                            Swal.fire({
+                                                                title: 'Apakah kamu yakin?',
+                                                                text: "kamu akan menghapus data ini!",
+                                                                icon: 'warning',
+                                                                showCancelButton: true,
+                                                                confirmButtonColor: '#3085d6',
+                                                                cancelButtonColor: '#d33',
+                                                                confirmButtonText: 'Ya, hapus!'
+                                                            }).then((result) => {
+                                                                if (result.isConfirmed) {
+                                                                    document.getElementById('formDelete' + id).submit();
+                                                                }
+                                                            })
+                                                        }
+                                                    </script>
+                                                    <a type="button" class="btn btn-warning" data-toggle="modal"
+                                                        data-target="#formUpdate{{ $item->id }}">
+                                                        <i class="fa fa-edit" title="Ubah Data User"></i>
+                                                    </a>
+                                                </td>
+                                            @endif
                                         </tr>
                                         <?php $i++; ?>
                                         @include('pages.kategori-obat.update')
