+++ resources/views/pages/pasien/index.blade.php
@@ -19,9 +19,11 @@
                 <div class="col-12">
                     <div class="card">
                         <div class="card-body">
-                            <a type="button" class="btn btn-primary" data-toggle="modal" data-target="#formCreate"><i
-                                    class="fa fa-plus"></i> Tambah Pasien</a>
-                            @include('pages.pasien.create')
+                            @if (request()->session()->get('user')['role'] === 'ADMIN')
+                                <a type="button" class="btn btn-primary" data-toggle="modal" data-target="#formCreate"><i
+                                        class="fa fa-plus"></i> Tambah Pasien</a>
+                                @include('pages.pasien.create')
+                            @endif
                             <table id="defaultTable" class="table table-bordered table-striped">
                                 <thead>
                                     <tr>
@@ -30,7 +32,9 @@
                                         <th>No Telp</th>
                                         <th>Tanggal Lahir</th>
                                         <th>Gol Darah</th>
-                                        <th>Aksi</th>
+                                        @if (request()->session()->get('user')['role'] === 'ADMIN')
+                                            <th>Aksi</th>
+                                        @endif
                                     </tr>
                                 </thead>
                                 <tbody>
@@ -44,40 +48,42 @@
                                             <td>{{ $item->no_telp }}</td>
                                             <td>{{ $item->tanggal_lahir }}</td>
                                             <td>{{ $item->golongan_darah }}</td>
-                                            <td>
-                                                <form id="formDelete{{ $item->id }}"
-                                                    action="{{ route('pasien.destroy', $item->id) }}" method="POST"
-                                                    class="d-inline">
-                                                    @csrf
-                                                    @method('delete')
-                                                    <a type="button" class="btn btn-danger"
-                                                        onclick="handleDelete({{ $item->id }})">
-                                                        <i class="fa fa-trash" title="Hapus Data User"></i>
-                                                    </a>
-                                                </form>
+                                            @if (request()->session()->get('user')['role'] === 'ADMIN')
+                                                <td>
+                                                    <form id="formDelete{{ $item->id }}"
+                                                        action="{{ route('pasien.destroy', $item->id) }}" method="POST"
+                                                        class="d-inline">
+                                                        @csrf
+                                                        @method('delete')
+                                                        <a type="button" class="btn btn-danger"
+                                                            onclick="handleDelete({{ $item->id }})">
+                                                            <i class="fa fa-trash" title="Hapus Data User"></i>
+                                                        </a>
+                                                    </form>
 
-                                                <script>
-                                                    function handleDelete(id) {
-                                                        Swal.fire({
-                                                            title: 'Apakah kamu yakin?',
-                                                            text: "kamu akan menghapus data ini!",
-                                                            icon: 'warning',
-                                                            showCancelButton: true,
-                                                            confirmButtonColor: '#3085d6',
-                                                            cancelButtonColor: '#d33',
-                                                            confirmButtonText: 'Ya, hapus!'
-                                                        }).then((result) => {
-                                                            if (result.isConfirmed) {
-                                                                document.getElementById('formDelete' + id).submit();
-                                                            }
-                                                        })
-                                                    }
-                                                </script>
-                                                <a type="button" class="btn btn-warning" data-toggle="modal"
-                                                    data-target="#formUpdate{{ $item->id }}">
-                                                    <i class="fa fa-edit" title="Ubah Data User"></i>
-                                                </a>
-                                            </td>
+                                                    <script>
+                                                        function handleDelete(id) {
+                                                            Swal.fire({
+                                                                title: 'Apakah kamu yakin?',
+                                                                text: "kamu akan menghapus data ini!",
+                                                                icon: 'warning',
+                                                                showCancelButton: true,
+                                                                confirmButtonColor: '#3085d6',
+                                                                cancelButtonColor: '#d33',
+                                                                confirmButtonText: 'Ya, hapus!'
+                                                            }).then((result) => {
+                                                                if (result.isConfirmed) {
+                                                                    document.getElementById('formDelete' + id).submit();
+                                                                }
+                                                            })
+                                                        }
+                                                    </script>
+                                                    <a type="button" class="btn btn-warning" data-toggle="modal"
+                                                        data-target="#formUpdate{{ $item->id }}">
+                                                        <i class="fa fa-edit" title="Ubah Data User"></i>
+                                                    </a>
+                                                </td>
+                                            @endif
                                         </tr>
                                         <?php $i++; ?>
                                         @include('pages.pasien.update')
